// src/components/commerce/QuantitySelector/QuantityVariants.d.ts
export declare const quantitySelectorVariants: (props?: ({
    size?: "sm" | "md" | "lg" | null | undefined;
    radius?: "full" | "none" | "sm" | "md" | "lg" | "xl" | null | undefined;
    disabled?: boolean | null | undefined;
} & import("class-variance-authority/types").ClassProp) | undefined) => string;
export declare const quantityInputVariants: (props?: ({
    size?: "sm" | "md" | "lg" | null | undefined;
} & import("class-variance-authority/types").ClassProp) | undefined) => string;